# api/handlers/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "semrel registry API",
    "version": "1.0.0",
    "description": "The semrel plugin registry. The public catalogue is readable without authentication; publishing and review require a GitHub session.\n\nBrowsers authenticate with an HttpOnly `semrel_session` cookie set by the OAuth callback. API clients (the semrel CLI, CI jobs) send `Authorization: Bearer <token>`.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "https://registry.semrel.io",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v1/plugins": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "summary": "List plugins",
        "description": "Anonymous callers see active plugins. An authenticated non-admin additionally sees their own plugins in every state; an admin may filter by status.",
        "security": [
          {},
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Category"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "Full-text and partial-identifier search. Results are ranked by relevance unless an explicit sort is given."
          },
          {
            "name": "compatibleWith",
            "in": "query",
            "schema": {
              "type": "string",
              "examples": [
                "0.27.1"
              ]
            },
            "description": "A concrete semrel core version. Returns only plugins whose latest installable release declares a compatibility range admitting it. Plugins that declare no range are included: the metadata is optional and most of the catalogue predates it."
          },
          {
            "name": "author",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Status"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "category",
                "created_at",
                "updated_at",
                "downloads",
                "views"
              ]
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of plugins.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      },
      "post": {
        "tags": [
          "publishing"
        ],
        "summary": "Create a plugin",
        "description": "Non-admin callers have the author forced to their own GitHub login.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/plugins/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/PluginRef"
        }
      ],
      "get": {
        "tags": [
          "catalogue"
        ],
        "summary": "Get one plugin",
        "security": [
          {},
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The plugin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginEnvelope"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "publishing"
        ],
        "summary": "Update a plugin",
        "description": "Non-admins may only update plugins they author, and may not reassign the author.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "publishing"
        ],
        "summary": "Soft-delete a plugin",
        "description": "Requires a typed confirmation matching the plugin reference.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginDeletionRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/schemas/plugins/{name}/{version}": {
      "get": {
        "tags": [
          "schemas"
        ],
        "summary": "Plugin configuration schema",
        "security": [],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON Schema."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "semrel_session",
        "description": "HttpOnly, SameSite=Lax session cookie set by the OAuth callback. Browser clients use this; mutating requests are additionally checked against the Origin header."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "For non-browser clients. The static ADMIN_TOKEN is accepted outside production only."
      },
      "webhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Hub-Signature-256",
        "description": "HMAC-SHA256 over the raw request body, keyed with WEBHOOK_SECRET, formatted as sha256=<hex>."
      }
    },
    "parameters": {
      "PluginRef": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Numeric id, plugin name, or alias.",
        "examples": {
          "byName": {
            "value": "provider-github"
          },
          "byId": {
            "value": "42"
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "The request was rejected by validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication is required, or the session has expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated, but not permitted. Publishers may only touch their own plugins.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "No such plugin or version.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Too many requests. X-RateLimit-Limit and X-RateLimit-Remaining describe the budget.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          },
          "X-RateLimit-Limit": {
            "schema": {
              "type": "string"
            }
          },
          "X-RateLimit-Remaining": {
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Category": {
        "type": "string",
        "enum": [
          "analyzer",
          "condition",
          "generator",
          "hook",
          "provider",
          "updater",
          "packager",
          "publisher"
        ]
      },
      "Status": {
        "type": "string",
        "enum": [
          "active",
          "pending",
          "rejected"
        ]
      },
...
      "PluginInput": {
        "type": "object",
        "required": [
          "name",
          "category"
        ],
        "properties": {
          "namespace": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "maxLength": 4000
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "repository": {
            "type": "string",
            "format": "uri"
          },
          "license": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PluginList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Plugin"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "PluginEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Plugin"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerAuth": []
    }
  ]
}